+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- New class `pydictdb.core.Query` to search objects in dictionary.
+- New module `pydictdb.core` to operate CRUD on dict data.
 - New module `pydictdb.storages` to read and write data in memory or file.